% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorias_pregunta_formato.R
\name{categorias_pregunta_formato}
\alias{categorias_pregunta_formato}
\title{Formato categorías respuestas para preguntas múltiples y categóricas}
\usage{
categorias_pregunta_formato(
pregunta,
diseño,
datos,
DB_Mult,
tipo_pregunta,
metricas = TRUE
)
}
\arguments{
\item{pregunta}{Nombre de la pregunta sobre la cual se desea obtener las categorías}

\item{diseño}{Diseño muestral que se ocupará según el tipo de pregunta}

\item{datos}{Conjunto de datos en formato .sav}

\item{DB_Mult}{Data frame con las preguntas múltiples}

\item{tipo_pregunta}{Tipo de pregunta: 'categorica', 'multiple', 'continua'}

\item{metricas}{Valor lógica que indica si se desean las categorías de las respuestas para las métricas media,ímite inferior, límite superior o para las métricas error estándar, varianza, coeficiente de variación y el efecto de diseño}
}
\value{
Vector con las categorías de las respuestas de la pregunta mecionada por el usuario, si métricas = TRUE se regresa un vector con las categorías y entre cada categoría se encuentran dos NA's,

si metricas = FALSE se regresa un vector con las categorías de las respuestas y entre cada categoría se encuentran 3 NA's
}
\description{
Crea un vector para las respuestas de las preguntas múltiples y categóricas.
}
\examples{
\dontrun{
# Carga de datos
dataset <- read.spss("data/BASE_CONACYT_260118.sav", to.data.frame = TRUE)
Lista_Preg <- read_xlsx("aux/Lista de Preguntas.xlsx",
  sheet = "Lista Preguntas"
)$Nombre \%>\% as.vector()
DB_Mult <- read_xlsx("aux/Lista de Preguntas.xlsx",
  sheet = "Múltiple"
) \%>\% as.data.frame()

# Diseño
disenio_mult <- disenio(id = c(CV_ESC, ID_DIAO), estrato = ESTRATO, pesos = Pondi1, reps = FALSE, datos = dataset)

categorias_pregunta_formato(pregunta = "P1", diseño = disenio_mult, datos = dataset, DB_Mult = DB_Mult, tipo_pregunta = "multiple", metricas = TRUE)
}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
