% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disenio.R
\name{disenio}
\alias{disenio}
\title{Crea un objeto con un diseño de muestreo}
\usage{
disenio(
id,
estrato,
pesos,
datos,
pps = "brewer",
varianza = "HT",
reps = TRUE,
metodo = "subbootstrap",
B = 500,
semilla = 1234
)
}
\arguments{
\item{id}{Variable que especifica el indicador para cada uno de los estratos}

\item{estrato}{Variable que especifica los estratos.}

\item{pesos}{Variables que especifica las ponderaciones.}

\item{datos}{Conjunto de datos en formato .sav}

\item{pps}{Valor que indica el tipo de aproximación que se desea utilizar para el muestreo.}

\item{varianza}{Valor que indica el estimador que se desea utilizar ("HT" para el estimador Horvitz-Thompson, "YG" para el estimador Yates-Grundy)}

\item{reps}{Valor lógico que indica si se desea realizar un muestreo con pesos replicados.}

\item{metodo}{Si reps = TRUE; tipo de replicación de pesos que se desea usar.}

\item{B}{Si reps = TRUE; número de replicaciones que desean realizar.}

\item{semilla}{Número utilizado para inicializar la generación números aleatorios.}
}
\value{
Objeto del tipo tbl_svy
}
\description{
Se crea un objeto con el diseño de muestreo especificado.
}
\examples{
\dontrun{
disenio(
  id = id_estrato, estrato = estrato, pesos = ponderador, datos = dataset,
  pps = "brewer", varianza = "HT", reps = TRUE, metodo = "subbootstrap",
  B = 50, semilla = 1234
)
}
}
\seealso{
\code{\link{as_survey_design}} \code{\link{as_survey_rep}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
