% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estilo_columnas.R
\name{estilo_columnas}
\alias{estilo_columnas}
\title{Divide las métricas de las tablas cruzadas}
\usage{
estilo_columnas(
tabla,
wb,
hojas =  c(3,4),
estilo = verticalStyle,
renglon = c(1,1),
tipo_pregunta = "categorica"
)
}
\arguments{
\item{tabla}{Tabla cruzada creada con la función tablas_cruzadas()}

\item{wb}{wb Workbook en el que aplicará el estilo de columnas}

\item{hojas}{Indicador de las hojas del workbook en las que se aplicará el estilo de columnas}

\item{estilo}{Un objeto de estilo devuelto por la función createStyle()}

\item{renglon}{Renglón donde se iniciará el formato}

\item{tipo_pregunta}{Tipo de pregunta: 'categorica', 'multiple', 'continua'}
}
\description{
Divide las métricas de las tablas cruzadas por categoría de respuestas en el caso de preguntas catégoricas y múltiples.
}
\examples{
\dontrun{
tabla_cruzada <- tablas_cruzadas(
  diseño = disenio_mult,
  pregunta = 'P1',
  dominio = 'Sexo',
  datos = dataset,
  DB_Mult = DB_Mult,
  tipo_pregunta = 'multiple'
  )

estilo_columnas(
  tabla = tabla_cruzada,
  wb = wb,
  hojas = c(3,4),
  estilo = verticalStyle,
  renglon = c(1,1),
  tipo_pregunta = 'categorica'
  )
}
}
\seealso{
\code{\link{createStyle}} \code{\link{addStyle}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
