% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estilo_dominios.R
\name{estilo_dominios}
\alias{estilo_dominios}
\title{Da formato a la columna de dominios de las tablas cruzadas en el workbook}
\usage{
estilo_dominios(
tabla,
wb,
columna = 1,
hojas = c(3,4),
dominios,
renglon = c(1,1),
estilo_dominios = horizontalStyle,
estilo_merge_dominios = bodyStyle,
tipo_pregunta
)
}
\arguments{
\item{tabla}{Lista con las tablas cruzadas formateadas}

\item{wb}{Workbook en el que aplicará el formato}

\item{columna}{Columna en la que se aplicará el formato}

\item{hojas}{Hojas del workbook a las que se aplicará el formato, una por tabla}

\item{dominios}{Vector con los dominios}

\item{renglon}{Renglón donde se iniciará el formato, uno por tipo de tabla, uno por tabla}

\item{estilo_dominios}{Estilo previamente definido con el borde en la parte inferior de la tabla}

\item{estilo_merge_dominios}{Estilo previamente definido con bordes en toda la tabla}

\item{tipo_pregunta}{Tipo de pregunta ("categorica", "multiple")}
}
\description{
Función para unir las celdas donde se encuentran los nombres de los dominios es las hojas solicitadas.
}
\examples{
\dontrun{
estilo_dominios(
  tabla = tabla_cruzada, wb = wb, columna = 1, hojas = c(3, 4),
  dominios = dominios, renglon = c(1, 1), estilo_dominios = horizontalStyle,
  estilo_merge_dominios = bodyStyle, tipo_pregunta = "categorica"
)
}
}
\seealso{
\code{\link{survey_mean}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
