% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatear_frecuencias_simples.R
\name{formatear_frecuencias_simples}
\alias{formatear_frecuencias_simples}
\title{Formatea tabla frecuencias simples}
\usage{
formatear_frecuencias_simples(
tabla,
tipo_pregunta = "categorica"
)
}
\arguments{
\item{tabla}{Tabla de frecuencias simples creada con la función frecuencias_simples()}

\item{tipo_pregunta}{Tipo de pregunta: 'categorica', 'multiple', 'continua'}
}
\value{
Lista de dos tibbles: en la primera tibble se encuentra el total estimado y las métricas media, límite inferior y superior;
en la segunda tibble se encuentra el total y las métricas error estándar, varianza, coeficiente de variación y el efecto de diseño (si es que DEFF = TRUE en la función frecuencias_simples)
}
\description{
Se formatea la tabla de frecuencias simples según tipo de pregunta
}
\examples{
\dontrun{
freq_simples <- frecuencias_simples(diseño = disenio_cat, datos = dataset, pregunta = 'P1',
 DB_Mult = DB_Mult, tipo_pregunta = 'categorica')

frecuencias_simples(tabla = freq_simples, tipo_pregunta = 'categorica')
}
}
\seealso{
\code{\link{survey_mean}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
