% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatear_tabla_cruzada.R
\name{formatear_tabla_cruzada}
\alias{formatear_tabla_cruzada}
\title{Formatea tabla cruzada}
\usage{
formatear_tabla_cruzada(
pregunta,
datos,
dominio,
tabla,
DB_Mult,
tipo_pregunta
)
}
\arguments{
\item{pregunta}{Nombre de la pregunta sobre la cual se desea obtener la tabla cruzada}

\item{datos}{Conjunto de datos en formato .sav}

\item{dominio}{Dominio sobre el cual se desea desagregar la tabla cruzada de cierta pregunta}

\item{tabla}{Tabla cruzada creada con la función tablas_cruzadas()}

\item{tipo_pregunta}{Tipo de pregunta: 'categorica', 'multiple', 'continua'}
}
\value{
Lista de dos tibbles: en la primera tibble se encuentra el total estimado y las métricas media, límite inferior y superior;
en la segunda tibble se encuentra el total y las métricas error estándar, varianza, coeficiente de variación y el efecto de diseño (si es que DEFF = TRUE)
}
\description{
Se formatea la tabla cruzada según tipo de pregunta y dominio
}
\examples{
\dontrun{
 dataset <- read.spss("data/BASE_CONACYT_260118.sav", to.data.frame = TRUE)
 Lista_Preg <- read_xlsx("aux/Lista de Preguntas.xlsx",
                       sheet = "Lista Preguntas")$Nombre \%>\% as.vector()
 DB_Mult <- read_xlsx("aux/Lista de Preguntas.xlsx",  sheet = "Múltiple") \%>\% as.data.frame()
 Lista_Cont <- read_xlsx("aux/Lista de Preguntas.xlsx",
  sheet = "Continuas")$VARIABLE \%>\% as.vector()
 Dominios <- read_xlsx("aux/Lista de Preguntas.xlsx", sheet = "Dominios")$Dominios \%>\% as.vector()

 disenio_mult <- disenio(id = c(CV_ESC, ID_DIAO), estrato = ESTRATO, pesos = Pondi1, reps=FALSE, datos = dataset)

 tc <-  tablas_cruzadas(diseño = disenio_mult, pregunta = 'P1', dominio = 'Sexo', datos = dataset,
 DB_Mult = DB_Mult, tipo_pregunta = 'multiple')

 formatear_tabla_cruzada(pregunta = 'P1', datos = dataset, dominio = 'Sexo', tabla = tc, DB_Mult = DB_Mult, tipo_pregunta = 'multiple')
}
}
\seealso{
\code{\link{survey_mean}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
