% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formato_categorias.R
\name{formato_categorias}
\alias{formato_categorias}
\title{Función formato categorías para preguntas categóricas y múltiples}
\usage{
formato_categorias(
tabla,
pregunta,
diseño,
datos,
DB_Mult,
wb,
renglon,
columna = 4,
hojas,
estilo_cuerpo,
tipo_pregunta
)
}
\arguments{
\item{tabla}{Tabla cruzada creada por la función total_general()}

\item{pregunta}{Nombre de la pregunta sobre la cual se creo la tabla}

\item{diseño}{Diseño muestral que se ocupará según el tipo de pregunta}

\item{datos}{Conjunto de datos en formato .sav}

\item{DB_Mult}{Data frame con las preguntas múltiples}

\item{wb}{Workbook de Excel que contiene al menos dos hojas}

\item{renglon}{Vector tamaño 2 especificando el número de renglon en el cual se desea empezar a escribir el vector de categorías}

\item{columna}{Columna en la cual se desea empezar a escribir las tablas. SIEMPRE EN LA CUARTA COLUMNA}

\item{hojas}{Vector de número de hojas en el cual se desea insertar los vectores}

\item{estilo_cuerpo}{Estilo que indica cómo se desea formatear el vector en las hojas de excel}

\item{tipo_pregunta}{Tipo de pregunta_ 'categorica', 'multiple', 'continua'}
}
\description{
El vector obtenido con la función categorias_pregunta_formato() se escribe en un workbook de Excel
}
\details{
Una vez escrito el vector colapsa celdas (3 o 4)
}
\examples{
\dontrun{
# Creación del workbook
wb <- createWorkbook()
addWorksheet(wb, "Tablas cruzadas")
addWorksheet(wb, "Tablas cruzadas (dispersión)")

# Estilos
bodyStyle <- createStyle(
  halign = "center", border = "TopBottomLeftRight",
  borderColour = "black", borderStyle = "thin",
  valign = "center", wrapText = TRUE
)

# Carga de datos
dataset <- read.spss("data/BASE_CONACYT_260118.sav", to.data.frame = TRUE)
Lista_Preg <- read_xlsx("aux/Lista de Preguntas.xlsx",
  sheet = "Lista Preguntas"
)$Nombre \%>\% as.vector()
DB_Mult <- read_xlsx("aux/Lista de Preguntas.xlsx", sheet = "Múltiple") \%>\% as.data.frame()
Lista_Cont <- read_xlsx("aux/Lista de Preguntas.xlsx",
  sheet = "Continuas"
)$VARIABLE \%>\% as.vector()
Dominios <- read_xlsx("aux/Lista de Preguntas.xlsx", sheet = "Dominios")$Dominios \%>\% as.vector()

disenio_mult <- disenio(id = c(CV_ESC, ID_DIAO), estrato = ESTRATO, pesos = Pondi1, reps = FALSE, datos = dataset)

total <- total_general(
  diseño = disenio_mult, pregunta = "P1", dominio = "General", datos = dataset,
  DB_Mult = DB_Mult, tipo_pregunta = "multiple"
)

formato_categorias(
  tabla = total, pregunta = "P1", diseño = disenio_mult, datos = dataset,
  DB_Mult = DB_Mult, wb = wb, renglon = c(1, 1), columna = 4, hojas = c(1, 2), estilo_cuerpo = bodyStyle, tipo_pregunta = "multiple"
)

openxlsx::openXL(wb)
}
 @import openxlsx
}
\seealso{
\code{\link{mergeCells}} \code{\link{addStyle}} \code{\link{createStyle}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
