% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formato_frecuencias_simples.R
\name{formato_frecuencias_simples}
\alias{formato_frecuencias_simples}
\title{Escribe frecuencias simples en dos hojas de Excel}
\usage{
formato_frecuencias_simples(
tabla,
wb,
hojas,
renglon,
columna,
estilo_encabezado,
estilo_horizontal,
estilo_total,
tipo_pregunta
)
}
\arguments{
\item{tabla}{lista de las tibbles creadas por la función formatear_frecuencias_simples}

\item{wb}{Workbook de Excel que contiene al menos dos hojas}

\item{hojas}{vector de número de hojas en el cual se desea insertar las tablas}

\item{renglon}{vetor tamaño 2 especificando el número de renglon en el cual se desea empezar a escribir la tabla 1 y tabla 2 respectivamente}

\item{columna}{columna en la cual se desea empezar a escribir las tablas}

\item{estilo_encabezado}{estilo el cual se desea usar para los nombres de las columnas}

\item{estilo_horizontal}{estilo último renglón horizontal}

\item{estilo_total}{estilo el cual se desea usar para la columna total}

\item{tipo_pregunta}{tipo de pregunta_ 'categorica', 'multiple', 'continua'}
}
\description{
Escribe la tabla de frecuencias simples formateadas en dos hojas de Excel
}
\details{
Ambas tablas se escribirán en la misma columna pero en diferentes hojas (especificadas en el parámetro hojas).

El estilo_total se recomienda crear un estilo con la función createStyle de openxlsx con el formato que se desea, por ejemplo "###,###,###.0"

El estilo_horizontal hace referencia al tipo de lineado horizontal se desea en el úntimo renglón de la tabla
}
\examples{
\dontrun{
# Creación del workbook
wb <- createWorkbook()
addWorksheet(wb, "Frecuencias simples")
addWorksheet(wb, "Frecuencias simples (dispersión)")

# Estilos
headerStyle <- createStyle(fontSize = 11, fontColour = "black", halign = "center",
border = "TopBottom", borderColour = "black",
borderStyle = c('thin', 'double'), textDecoration = 'bold')

totalStyle <-  createStyle(numFmt = "###,###,###.0")

horizontalStyle <- createStyle(border = "bottom", borderColour = "black", borderStyle = 'thin', valign = 'center')

freq_simples <- frecuencias_simples(diseño = disenio_cat, datos = dataset, pregunta = 'P1',
 DB_Mult = DB_Mult, tipo_pregunta = 'categorica')

freq_simples_format <- frecuencias_simples(tabla = freq_simples, tipo_pregunta = 'categorica')

formato_frecuencias_simples(tabla = freq_simples_format , wb = wb, hojas = c(1,2),
renglon = c(1,1), columna = 1 , estilo_encabezado = headerStyle ,
estilo_horizontal = horizontalStyle , estilo_total = totalStyle, tipo_pregunta = 'categorica')
}
}
\seealso{
\code{\link{writeData}} \code{\link{createStyle}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
