% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formato_tablas_cruzadas.R
\name{formato_tablas_cruzadas}
\alias{formato_tablas_cruzadas}
\title{Escribe y da formato a la tablas cruzadas en el workbook}
\usage{
formato_tablas_cruzadas(
tabla,
wb,
renglon = c(1,1),
columna = 1,
hojas = c(3,4),
estilo_encabezado = headerStyle,
estilo_total = totalStyle
)
}
\arguments{
\item{tabla}{Lista con las tablas cruzadas formateadas}

\item{wb}{Workbook en el que escribirán las tablas}

\item{renglon}{Renglón donde iniciará la tabla, un renglón por tipo de tabla cruzada}

\item{columna}{Columna en la que se iniciará la tabla}

\item{hojas}{Hojas del workbook en donde se ecribirán las tablas}

\item{estilo_encabezado}{Estilo para el ancabezado de la tabla}

\item{estilo_total}{Estilo para el resto de la tabla}
}
\description{
Función para escribir y dar formato a las tablas cruzadas en el workbook
}
\examples{
\dontrun{
# Estilos
headerStyle <- createStyle(
  fontSize = 11, fontColour = "black", halign = "center",
  border = "TopBottom", borderColour = "black",
  borderStyle = c("thin", "double"), textDecoration = "bold"
)

totalStyle <- createStyle(numFmt = "###,###,###.0")
formato_tablas_cruzadas(
  tabla = tabla_cruzada, wb = wb, renglon = c(1, 1),
  columna = 1, hojas = c(3, 4), estilo_encabezado = headerStyle,
  estilo_total = totalStyle
)
}
}
\seealso{
\code{\link{survey_mean}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
