% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frecuencias_simples_excel.R
\name{frecuencias_simples_excel}
\alias{frecuencias_simples_excel}
\title{Función frecuencias simples en Excel}
\usage{
frecuencias_simples_excel(
pregunta,
num_pregunta,
datos,
DB_Mult,
lista_preguntas,
diseño,
wb,
renglon = c(1,1),
columna = 1,
hojas = c(1,2),
tipo_pregunta = 'categorica',
fuente,
pie_tabla,
logo_path,
organismo_participacion,
estilo_encabezado = headerStyle,
estilo_horizontal = horizontalStyle,
estilo_total = totalStyle
 )
}
\arguments{
\item{pregunta}{Nombre de la pregunta sobre la cual se desea obtener frecuencias simples e incluirlas en un workbook de Excel}

\item{num_pregunta}{Número de pregunta}

\item{datos}{Conjunto de datos en formato .sav}

\item{DB_Mult}{Data frame con las preguntas múltiples}

\item{lista_preguntas}{Data frame que contiene los títulos de las pregunta}

\item{diseño}{Diseño muestral que se ocupará según el tipo de pregunta}

\item{wb}{Workbook de Excel que contiene al menos dos hojas}

\item{renglon}{Vector tamaño 2 especificando el número de renglon en el cual se desea empezar a escribir la tabla 1 y tabla 2 respectivamente}

\item{columna}{Columna en la cual se desea empezar a escribir las tablas}

\item{hojas}{Vector de número de hojas en el cual se desea insertar las tablas}

\item{tipo_pregunta}{Tipo de pregunta_ 'categorica', 'multiple', 'continua'}

\item{fuente}{Nombre del proyecto}

\item{pie_tabla}{Pie de tabla}

\item{organismo_participacion}{Organismos que participaron en el proyecto, por ejemplo, 'Ciudadanía Mexicana'}

\item{logo_path}{Path del logo de la UNAM}

\item{estilo_encabezado}{estilo el cual se desea usar para los nombres de las columnas}

\item{estilo_horizontal}{estilo último renglón horizontal}

\item{estilo_total}{estilo el cual se desea usar para la columna total}
}
\description{
Escribe los títulos 'Frecuencias simples' y 'Frecuencias simples (dispersión)', logo indicado, título de la pregunta, tabla de frecuencias simples formateada y pie de tabla en las hojas y renglones mencionados por el usuario
}
\details{
Esta función envuelve todas las funciones creadas para obtener las frecuencias simples, por lo que esta función es la única que se deberá llamar para crear las frecuencias simples de las preguntas deseadas e insertarlas en ciertas hojas de Excel

Es necesario crear al menos dos hojas de excel con la función addWorksheet de la paquetería openxlsx

El estilo_total se recomienda crear un estilo con la función createStyle de openxlsx con el formato que se desea, por ejemplo "###,###,###.0"

El estilo_horizontal hace referencia al tipo de lineado horizontal se desea en el úntimo renglón de la tabla
}
\examples{
\dontrun{
# Creación del workbook
wb <- createWorkbook()
addWorksheet(wb, "Frecuencias simples")
addWorksheet(wb, "Frecuencias simples (dispersión)")

# Estilos
headerStyle <- createStyle(fontSize = 11, fontColour = "black", halign = "center",
border = "TopBottom", borderColour = "black",
borderStyle = c('thin', 'double'), textDecoration = 'bold')
totalStyle <-  createStyle(numFmt = "###,###,###.0")
horizontalStyle <- createStyle(border = "bottom",
borderColour = "black", borderStyle = 'thin', valign = 'center')

# Carga de datos
 dataset <- read.spss("data/BASE_CONACYT_260118.sav", to.data.frame = TRUE)
 Lista_Preg <- read_xlsx("aux/Lista de Preguntas.xlsx",
 sheet = "Lista Preguntas")$Nombre \%>\% as.vector()
  DB_Mult <- read_xlsx("aux/Lista de Preguntas.xlsx",
  sheet = "Múltiple") \%>\% as.data.frame()

#Diseño
 disenio_mult <- disenio(id = c(CV_ESC, ID_DIAO), estrato = ESTRATO, pesos = Pondi1, reps=FALSE, datos = dataset)

frecuencias_simples_excel(pregunta = 'P1',
num_pregunta = 1,
datos = dataset,
DB_Mult = DB_Mult,
lista_preguntas = Lista_Preg,
diseño = disenio_mult,
wb = wb,
renglon = c(1,1),
columna = 1,
hojas = c(1,2),
tipo_pregunta = 'multiple',
fuente =  'Conacyt 2018',
pie_tabla = 'Conacyt 2018',
organismo_participacion = 'Ciudadanía mexicana',
logo_path = 'img/logo_unam.png',
estilo_encabezado = headerStyle,
estilo_horizontal = horizontalStyle,
estilo_total = totalStyle
)
}
}
\seealso{
\code{\link{writeData}} \code{\link{createStyle}}  \code{\link{setRowHeights}} \code{\link{insertImage}} \code{\link{mergeCells}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
