% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_cruzada_total.R
\name{tabla_cruzada_total}
\alias{tabla_cruzada_total}
\title{Función tabla cruzada total y por dominios}
\usage{
tabla_cruzada_total(
diseño,
pregunta,
datos,
dominios,
tipo_pregunta
)
}
\arguments{
\item{diseño}{Diseño muestral que se ocupará según el tipo de pregunta}

\item{pregunta}{Nombre de la pregunta sobre la cual se desea obtener la tabla cruzada total,}

\item{datos}{Conjunto de datos en formato .sav}

\item{dominios}{Vector el cual contiene los nombres de los dominios sobre los cuales se desean obtener las tablas cruzadas}

\item{tipo_pregunta}{Tipo de pregunta: 'categorica', 'multiple', 'continua'}
}
\value{
Lista de dos tibbles: en la primera tibble se encuentra el total estimado y las métricas media, límite inferior y superior;
en la segunda tibble se encuentra el total y las métricas error estándar, varianza, coeficiente de variación y el efecto de diseño (si es que DEFF = TRUE en la función tablas_cruzadas)
}
\description{
Esta función une la tabla cruzada general y las tablas cruzadas por dominios generadas por las funciones total_general() y tablas_cruzadas()
}
\examples{
\dontrun{
 # Lectura de datos
 dataset <- read.spss("data/BASE_CONACYT_260118.sav", to.data.frame = TRUE)
 Lista_Preg <- read_xlsx("aux/Lista de Preguntas.xlsx",
                       sheet = "Lista Preguntas")$Nombre \%>\% as.vector()
 DB_Mult <- read_xlsx("aux/Lista de Preguntas.xlsx",  sheet = "Múltiple") \%>\% as.data.frame()
 Lista_Cont <- read_xlsx("aux/Lista de Preguntas.xlsx",
  sheet = "Continuas")$VARIABLE \%>\% as.vector()
 Dominios <- read_xlsx("aux/Lista de Preguntas.xlsx", sheet = "Dominios")$Dominios \%>\% as.vector()

# Diseño
 disenio_mult <- disenio(id = c(CV_ESC, ID_DIAO), estrato = ESTRATO, pesos = Pondi1, reps=FALSE, datos = dataset)

 tabla_cruzada_total (diseño = disenio_mult,  pregunta = 'P1', dominios = Dominios, datos = dataset,
 DB_Mult = DB_Mult, tipo_pregunta = 'multiple')
}
}
\seealso{
\code{\link{survey_mean}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
