% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas_cruzadas.R
\name{tablas_cruzadas}
\alias{tablas_cruzadas}
\title{Función tablas cruzadas según tipo de pregunta y dominio}
\usage{
tablas_cruzadas(diseño,
pregunta,
dominio,
datos,
DB_Mult,
na.rm,
vartype = c("ci","se","var","cv"),
cuantiles = c(0,0.25, 0.5, 0.75,1),
significancia = 0.95,
proporcion = FALSE,
metodo_prop = 'likelihood',
DEFF = TRUE,
tipo_pregunta = 'categorica'
)
}
\arguments{
\item{diseño}{Diseño muestral que se ocupará según el tipo de pregunta}

\item{pregunta}{Pregunta de la cual se quieren obtener las frecuencias simples, por ejemplo, 'P_1'}

\item{dominio}{Nombre del dominio del cual se desea obtener la tabla cruzada}

\item{datos}{Conjunto de datos en formato .sav}

\item{DB_Mult}{Data frame con las preguntas múltiples}

\item{na.rm}{Valor lógico que indica si se deben de omitir valores faltantes}

\item{vartype}{Métricas de variabilidad: error estándar ("se"), intervalo de confianza ("ci"), varianza ("var") o coeficiente de variación ("cv")}

\item{cuantiles}{Vector de cuantiles a calcular}

\item{significancia}{Nivel de confianza: 0.95 por default}

\item{proporcion}{Valor lógico que indica si se desen usar métodos para calcular la proporción que puede tener intervalos de confianza más precisos cerca de 0 y 1}

\item{metodo_prop}{Si proporcion = TRUE; tipo de método de proporción que se desea usar: "logit", "likelihood", "asin", "beta", "mean"}

\item{DEFF}{Valor lógico que indica si se desea calcular el efecto de diseño}

\item{tipo_pregunta}{Tipo de pregunta: 'categorica', 'multiple', 'continua'}
}
\value{
Tabla tipo tibble con las estadísticas especificadas en el parámetro estadisticas por respuestas pertenecientes a la pregunta y al dominio especificados
}
\description{
Se crea la tabla cruzada según dominio tipo de pregunta (categórica, múltiple o continua)
}
\examples{
\dontrun{
 dataset <- read.spss("data/BASE_CONACYT_260118.sav", to.data.frame = TRUE)
 Lista_Preg <- read_xlsx("aux/Lista de Preguntas.xlsx",
                       sheet = "Lista Preguntas")$Nombre \%>\% as.vector()
 DB_Mult <- read_xlsx("aux/Lista de Preguntas.xlsx",  sheet = "Múltiple") \%>\% as.data.frame()
 Lista_Cont <- read_xlsx("aux/Lista de Preguntas.xlsx",
  sheet = "Continuas")$VARIABLE \%>\% as.vector()
 Dominios <- read_xlsx("aux/Lista de Preguntas.xlsx", sheet = "Dominios")$Dominios \%>\% as.vector()

 disenio_mult <- disenio(id = c(CV_ESC, ID_DIAO), estrato = ESTRATO, pesos = Pondi1, reps=FALSE, datos = dataset)

 tablas_cruzadas(diseño = disenio_mult, pregunta = 'P1', dominio = 'Sexo', datos = dataset,
 DB_Mult = DB_Mult, tipo_pregunta = 'multiple')
 }
 @import srvyr
 @import dplyr
 @import tidyr
 @import tibble
 @rawNamespace import(caret, except = lift)
 @export
}
\seealso{
\code{\link{survey_mean}} \code{\link{srvyr::group_by}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
