% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_general.R
\name{total_general}
\alias{total_general}
\title{Función total nacional}
\usage{
total_general(
diseño,
pregunta,
DB_Mult,
datos,
dominio = 'General',
tipo_pregunta,
na.rm = TRUE,
vartype = c("se","ci","cv", "var"),
cuantiles =  c(0,0.25, 0.5, 0.75,1),
significancia = 0.95,
proporcion = FALSE,
metodo_prop = "likelihood",
DEFF = TRUE
)
}
\arguments{
\item{diseño}{Diseño muestral que se ocupará según el tipo de pregunta}

\item{pregunta}{Nombre de la pregunta sobre la cual se desea obtener la tabla general}

\item{DB_Mult}{Data frame con las preguntas múltiples}

\item{datos}{Conjunto de datos en formato .sav}

\item{dominio}{Nombre al cual se desea nombrar al total estimado, por ejemplo 'General', 'Total Nacional', etc.}

\item{tipo_pregunta}{Tipo de pregunta: 'categorica', 'multiple', 'continua'}

\item{na.rm}{Valor lógico que indica si se deben de omitir valores faltantes}

\item{vartype}{Métricas de variabilidad: error estándar ("se"), intervalo de confianza ("ci"), varianza ("var") o coeficiente de variación ("cv")}

\item{cuantiles}{Vector de cuantiles a calcular}

\item{significancia}{Nivel de confianza: 0.95 por default}

\item{proporcion}{Valor lógico que indica si se desen usar métodos para calcular la proporción que puede tener intervalos de confianza más precisos cerca de 0 y 1}

\item{metodo_prop}{Si proporcion = TRUE; tipo de método de proporción que se desea usar: "logit", "likelihood", "asin", "beta", "mean"}

\item{DEFF}{Valor lógico que indica si se desea calcular el efecto de diseño}
}
\value{
Lista de dos tibbles: en la primera tibble se encuentra el total estimado y las métricas media, límite inferior y superior;
en la segunda tibble se encuentra el total y las métricas error estándar, varianza, coeficiente de variación y el efecto de diseño (si es que DEFF = TRUE)
}
\description{
Esta función transforma la tabla de frecuencias simples de cierta pregunta especificada por el usuario en un formato similar al de tablas cruzadas (1 rengón)
}
\examples{
\dontrun{
# Lectura de datos
 dataset <- read.spss("data/BASE_CONACYT_260118.sav", to.data.frame = TRUE)
 Lista_Preg <- read_xlsx("aux/Lista de Preguntas.xlsx",
                       sheet = "Lista Preguntas")$Nombre \%>\% as.vector()
 DB_Mult <- read_xlsx("aux/Lista de Preguntas.xlsx",  sheet = "Múltiple") \%>\% as.data.frame()
 Lista_Cont <- read_xlsx("aux/Lista de Preguntas.xlsx",
  sheet = "Continuas")$VARIABLE \%>\% as.vector()
 Dominios <- read_xlsx("aux/Lista de Preguntas.xlsx", sheet = "Dominios")$Dominios \%>\% as.vector()

# Diseño
 disenio_mult <- disenio(id = c(CV_ESC, ID_DIAO), estrato = ESTRATO, pesos = Pondi1, reps=FALSE, datos = dataset)

 total_general (diseño = disenio_mult,  pregunta = 'P1', dominio = 'General', datos = dataset,
 DB_Mult = DB_Mult, tipo_pregunta = 'multiple')
}
}
\seealso{
\code{\link{survey_mean}}
}
\author{
Bringas Arturo, Rosales Cinthia, Salgado Iván, Torres Ana
}
